Create manual_installation_production.md
I think it's this. Please review to be sure that I am not missing anything.
I was a bit confused with the directory links in the installer instructions, because in unix you don't create directories to then link them, since the command `ln -s` already creates the linked directory. But I think it's this. Please review and amend it accordingly